move #10 - api 폴더 병합, 코드 이동
